Broke tight coupling between the CategoryVM and the CategoriesListViewModel.
Added debgger friendly code so we get better errors when a view model cannot be resolved.

Created an app level nav service.  The idea here is that it will have a more semantic api, "Show This" "Shwo that".  It will know how to create the pages and show the content.